+++ src/index.css
@@ -206,11 +206,12 @@ h3 {
 h4 {
   color: var(--coalBlack);
   font-family: 'Kanit', sans-serif;
-  font-size: 1.5rem;
+  font-size: 1.7rem;
   font-weight: 400;
   padding: 10px;
-  margin-bottom: 10px;
-  margin-top: 10px;
+  margin-bottom: 0;
+  margin-top: 0;
+  text-align: center;
 }
 
 .all-projects-container {
@@ -283,6 +284,15 @@ img {
 .project-info-card p {
   line-height: 1.4rem;
   margin-bottom: 10px;
+}
+
+.contact-box {
+  margin-bottom: 50px;
+}
+
+.contact-box p {
+  margin: 0;
+  text-align: center;
 }
 
 .form {
